Fixes and speed mapping enhancement
This driver was not able to actually set all of the settings on the fan controller for me.  I could not change the indicator timeout, preset level, etc.  This seems to have been caused by the zwave configuration set commands not explicitly specifying the size of 1.  When I added that they all started working.  Possibly relevant is that my controller is firmware 1.8.

Also I've added a way to customize the mapping of the five Hubitat speeds on to the four supported by the controller.  You can choose in the device preferences how to map them, and all the various setting of speed and/or level is handled based on the chosen mapping.

Misc fixes:
* properly setting the FanControl capability's "supportedFanModes" based on the chosen speed mapping
* added the FanControl capability's "cycleSpeed" method that cycles through the supported modes similar to pulling the chain on a ceiling fan
* tweaked debug logging slightly to include device name in the log messages